Remove don't bother checks
With the removal of the test runner from the art makefiles,
as well as fixing an unnecessary repetive use of $(shell),
it only takes 300ms to always parse art/Android.mk with tests,
and a 6MB increase in build.ninja (1.5%).  Remove art_test_bother,
so that makefiles don't have to be re-read when build art tests.
This will also cause the art tests to build in make checkbuild,
which hasn't been visible in MAKECMDGOALS since the switch to
kati.

art_dont_bother is used for clean-oat to avoid parsing all of the
art Android.mk files, but in practice is more likely to slow things
because it will still need to reread the much larger
out/soong/Android.mk.  m clean-oat without art_dont_bother is likely
not going to need to regen at all.

Bug: 26275726
Test: m checkbuild
Test: m clean-oat-host
Test: m test-art-host-run-test-dependencies
Change-Id: I6573fd7ab54a71f90c8bb3bb43a3d872552eeb6d