Revert^4 "Prevent UAF issues caused by static destructors"
The test 1944 was not very robust and could spuriously fail for
several reasons. Most notably it did not actually check that the exit
value was what we expected and it could cause linker warning messages
to be printed. The first we solved by checking for the exit value
directly. The second cannot be simply solved so we simply filter these
messages out.

We needed to fix a bug with run-test that caused it to quash the exit
status of the dalvikvm invoke and fix several run-tests that relied on
this behavior. We did this by using a run script to add the exit value
to the test output and squash the exit value.

This reverts commit 70f4415d0485f67922323557f2657b009f20696d.

Reason for revert: Made test more robust.

Test: ./test.py --host -j50
Bug: 69591477
Change-Id: I8580973e16d63583e0cb80825892b4604c2591b6