Update patch set 2
Patch Set 2:

This fails make checkbuild on aosp_x86-eng because valgrind is not supported on x86 with SSE4, which is part of our default flags.  Fixing the dependencies on target valgrind is easy, but the valgrind makefiles are a mess and also disable the host valgrind when the target is x86.

Patch-set: 2
Label: Presubmit-Ready=0