WIP: Add a way to have compressed odex/vdex in ZIP
Add logic for extracting dex files out of the vdex files in ZIP files
this enables ART to load and run APKs that have stripped dex but
contain a vdex.

Also added a dex2oat option for extracting the oat and vdex out of
ZIP files. If this option is passed in, dex2oat does no compilation
and instead extracts to the expected outputs.

Added unit test.

Test: test-art-host
Bug: 70934104

Change-Id: Ibf72a784ee993313a659e04f579ecb7e4fa15e01