Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> (1 comment)

This is asking me to remove undefined behavior from all of this code. For example, CheckJNI::GetField assigns 0 to JniValueType.J assuming it clears the float/double/object/.. variants of JniValueType. I'd prefer this CL not to become a UB clean-up CL, as its ambition is already quite bold and roll-back may be necessary.
To clean up the code the type JniValueType and the format strings are best removed. They are better replaced with variadic templates. JniValueType is a Dalvik inheritance and so its a shame that it wasn't cleaned up when initially brought over. I can look into that but I'd rather not do it here - unless you have a specific small issue you think memcpy can cover. Thanks.

Patch-set: 3
Reviewer: Gerrit User 1010118 <1010118@85c56323-6fa9-3386-8a01-6480fb634889>