Fix ThreadStress run-script's exit status.
Script test/004-ThreadStress/run executes the ThreadStress test
twice: the first time as a normal run, the second time in
locks-only mode with stack-dump lock profiling. However, if the
first run failed but the second did not, the script would return
an exit status of 0, which would hide the former failure. This
change ensure we return the first failing exit status, if any.

Test: art/test/testrunner/testrunner.py -t 004-ThreadStress
Change-Id: I5e4e4cc7d9311fe15637ea2f5248a0e9f2432d61