Update patch set 2
Patch Set 2:

(2 comments)

Hi Hans, thanks for looking at this. Lock-free multi-threading is always tricky so I apprecitate another set of eyes.

Thanks for the link.  I didn't know this algorithm was well known and had a name.  I renamed it in the code to match.

Context: This is to be used from simpleperf and from debug-malloc. In case of simpleperf, that means the reader is executed in different process altogether. This makes taking a traditional lock difficult. Furthermore even if we had cross-process mutex (do we?), it would feel kind of weird for profiler to be taking a lock in the process it is profiling (or potentially in all procresses in the system).  If a process crashes while the lock it taken, it will be never unlocked, so thombstones would be impossible.  Debug-malloc could potentially take a lock, but I would rather have one interface and one code-path if possible.

Reader API: The memory is going to be read by libunwind using ptrace and process_vm_readv to cross the process boundary, and it has its own wrappers around those. That would make providing API for reading tricky.  Furthermore, libunwind can not depend on libart, so it is unclear where the code would live.  Libunwind is the only reader of the data, so I am happy to just keep the code there.

Consistency: I don't see any problem with reading arbitrarily inconsistent data. Libunwind will make copy of the data, and it will not use it in any way until the final check verifies that the copy was successful. It doesn't even assume that the data will stay mmaped. Also, the goal is that libunwind should not crash on invalid data anyway, so if anything goes wrong, we should, at worst, get a missed profile sample.

Patch-set: 2
Label: Presubmit-Ready=0