Add support for knownfailure.json type field
This lets testrunner.py know about flaky tests so it can retry them.

There are some tests that are slightly flaky for various reasons. This
allows testrunner.py to be told about flaky tests and try re-running
them when they fail. A test can be marked as flaky by adding an entry
to knownfailures.json with:

{
  ...
  "type": "flake",
  ...
}

A flaky test is considered passing if it succeeds at least once. The
number of times a flaky test is retried be changed by using the
--retry-flakes flag. By default we run with --retry-flakes=1 meaning
we will retry each flaky test once.

It also lets one note that a failure is expected because some part of
the configuration is not supported by using the 'unsupported' type.

Bug: 73871735
Bug: 73177368
Test: ./test.py --host -j50
Test: ./test/testrunner/testrunner.py --host --jvmti-stress -j50
Test: Modify a test to be flaky and run
Change-Id: I5ad8faec52ce2aaf513e1fe6c602641ec63999fd