Update patch set 4
Patch Set 4:

(3 comments)

> Consistency: I don't see any problem with reading arbitrarily inconsistent data. Libunwind will make copy of the data, and it will not use it in any way until the final check verifies that the copy was successful. It doesn't even assume that the data will stay mmaped. Also, the goal is that libunwind should not crash on invalid data anyway, so if anything goes wrong, we should, at worst, get a missed profile sample.

There is a linked list involved here, right? Consider the following scenario:

1) Reader starts traversing the linked list. Gets suspended at node A.
2) Writer increments the seqlock; removes A; increments seqlock again; deletes A;
3) writer reallocates what used to be A to a different kind of object. What used to be the next pointer gets set to 42.
4) Reader resumes, reads and dereferences A.next. Reader crashes.

What prevents this?

In general seqlocks tend to be used with really simple data structures, where the reader critical section essentially consists of copying out a bunch of data without ever looking at it. At a minimum, you otherwise need to have the reader recheck validity before dereferencing a pointer.

But that's still not correct because the data structure itself has non-atomic fields. Thus the reading code is violating the compiler's assumptions about data not changing out from under it, and you're thus prone to the compiler "breaking" your code. Consider code in the reading thread:

unsigned tmp_x = x; // x protected by seq_lock
if (tmp_x < 2) {
  ...
  a:
  switch(tmp_x) {
    case 0: ...; break;
    case 1: ...; break;
  }
}

If x and tmp_x are not changed in any of the omitted code, the compiler is allowed to

1) omit the range check for tmp_x in the switch. It knows it's less than 2.
2) reuse the register holding tmp_x, and reload tmp_x from x at a, if it needs more registers.

If x was concurrently changed to 3 between the initial test and a, this can result in all sorts of weirdness, up to and including a wild branch due to an out-of-bounds access to a branch table.

There should really only be atomic accesses in a seqlock critical section.

We do have process-shared mutexes and rwlocks. yabinc@ did a lot of recent bionic work in that general area. They also have timed lock functions, so you can possibly, sort of, maybe recover from a process that crashed while holding one. And that's probably the best you can hope for under any of these schemes, since you're accessing permanently inconsistent data.

I agree it's weird for these monitoring tools to acquire locks, or otherwise affect the monitored processes actions. But IIUC, that's more or less impossible to avoid here; it's just not safe to have the monitored process deallocate data that's still being read by the monitoring process. It seems to me that the monitoring process has to make it's presence known somehow. And to some extent this seems an artifact caused by the monitored process also being the guardian of the metainformation.

Patch-set: 4
Reviewer: Gerrit User 1042828 <1042828@85c56323-6fa9-3386-8a01-6480fb634889>