Update patch set 4
Patch Set 4:

> 4) Reader resumes, reads and dereferences A.next. Reader crashes.

This will be 4) Readers resumes, reads seqlock and notices it changed.  It discards the copy of the node without ever using its data for any program logic, and it restarts the whole algorithm with new linked list head.

This is also the reason why the "switch" example can not crash.  The reader will not use the data directly, the 'x' will be read from the copy (only after we are sure we managed to create consistet copy).

This is also the explanation for the final fence.  Without it, we would have "4) Reader finishes copy of the node; checks seqlock; it matches the starting seqlock, so it thinks the data is consistent; however the entry could have been freed and reused, and the writes could have been reordered before the seqlock update.  Hence the final write-write fence is needed to prevent that".

In any case, thank you for trying to break the alogrithm.  Do you have any other ideas how the reader could fail?

Patch-set: 4
Reviewer: Gerrit User 1065473 <1065473@85c56323-6fa9-3386-8a01-6480fb634889>