Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> > 4) Reader resumes, reads and dereferences A.next. Reader crashes.
> 
> This will be 4) Readers resumes, reads seqlock and notices it changed.  It discards the copy of the node without ever using its data for any program logic, and it restarts the whole algorithm with new linked list head.

I think you can make that work, assuming a draconian restriction on the reader. Basically the seqlock needs to be rechecked before using any of the data that was read. All you can do in a single reader critical section is read a bunch of data from previously known addresses. But that restriction would need to be included in the comment. You cannot, dereference any of the pointers you read without rechecking the seqlock. That's not an obvious restriction

> 
> This is also the reason why the "switch" example can not crash.  The reader will not use the data directly, the 'x' will be read from the copy (only after we are sure we managed to create consistet copy).

The C++ standard is clear that this is undefined behavior, and thus may crash. We're well into second-guessing the effect of compiler optimizations here, which is never a good place to be.

Going there anyway, let's assume that the seqlock release is right after the load of x into tmp_x. If the switch statement contains only accesses to local memory, the compiler is entirely justified in moving it into the seqlock critical section, at which point we have the same problem again. Fences don't order local accesses.

To make this kind of, sort of work, you'd have to tell the programmer not to even use the data purely locally before the end of of the critical section. And that fundamentally doesn't make sense because that ordering doesn't actually change the official semantics of the code, and is thus a really hacky restriction.

The only real way out of this it to use memory_order_relaxed atomic accesses to the data inside the seqlock critical section. That may be OK for performance, since memory_order_relaxed accesses generally compile to just load/stores. But it's not going to look beautiful.

> 
> This is also the explanation for the final fence.  Without it, we would have "4) Reader finishes copy of the node; checks seqlock; it matches the starting seqlock, so it thinks the data is consistent; however the entry could have been freed and reused, and the writes could have been reordered before the seqlock update.  Hence the final write-write fence is needed to prevent that".

So I think you're saying we have

Writer:
W1) Acquire seqlock for writing, reading S, writing S+1
W1.5) release-fence
W2) Remove linked list entry A
W3) Release seqlock, writing sequence number S+2
W4) Clobber A

Reader:
R1) Read seqlock for reading, reading S
R2) Read deallocated A
R2.5) acquire-fence
R3) Reread seqlock, again getting S

Assuming R2 reads an atomic location written by W2, we have W1.5 synchronizes with R2.5 (32.9p2) and hence W1 happens-before R3. (Fences officially only affect atomic operations, since races on other accesses are undefined.) This is a contradiction, since the read saw S.

This doesn't rely on the extra fence. It does assume all accesses are atomic, because without that the standard says that nothing works at all.

It would really help to have a more accurate characterization of reader code.

Patch-set: 4
Reviewer: Gerrit User 1042828 <1042828@85c56323-6fa9-3386-8a01-6480fb634889>