Update patch set 4
Patch Set 4: Code-Review-2

While I agree their is value in marking tests as flakes, I fail to understand what we're trying to gain by re-trying. Re-trying for me only has the potential to load even more our bots for little value. (and re-trying could be down to the test itself, it doesn't need to be plumbed in the infra).

On the other hand, I do see value in having the test run and ensure it doesn't crash, so we don't report a pass/fail, but only report a crash. Is that what you're trying to accomplish?

Patch-set: 4
Reviewer: Gerrit User 1038443 <1038443@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-2