Update patch set 4
Patch Set 4: Presubmit-Ready+1

> Patch Set 4: Code-Review-2
> 
> While I agree there is value in marking tests as flakes, I fail to understand what we're trying to gain by re-trying. Re-trying for me only has the potential to load even more our bots for little value. (and re-trying could be down to the test itself, it doesn't need to be plumbed in the infra).

The increased test-infra load will be minimal. It should basically be just the same as adding a single additional test (in the rare case that it even runs).

The added value is that we currently have several tests that we don't run due to (very rare) flakes. Of the tests I mark as flaky in go/aog/632215 all of them have single-digit percent flake rates but currently we don't run them. One of them is never run at all under any configuration (080-oom-fragmentation). By allowing a single retry we allow tests to continue to run and to get results.

I don't want to put the retry in the test itself since that will make even worse the huge number of test run-configurations. Furthermore most tests are only flakey under specific variants which are most easily checked in the testrunner. By putting it in the testrunner knownfailures.json it is also easier to remove when the flakiness is fixed and more visible as long as it's there.

> On the other hand, I do see value in having the test run and ensure it doesn't crash, so we don't report a pass/fail, but only report a crash. Is that what you're trying to accomplish?

This would be very difficult to do because of how we run the tests (5 layers of shell scripts) and would be insufficient for our purposes anyway.

Of the tests I mark as flaky in go/aog/632215 5 of them fail due to a rare KI race during shutdown with daemon threads (b/26483935, b/73871735). In those cases the slight change in performance from redefine-stress changes the flake rate from ~0.2% to ~0.8% from my tests. This seems to be just barely enough to (between the 5 tests with this problem) cause somewhat frequent failures when enabled. Fixing this in the 'right' way would almost certainly involve slowing down thread suspension (by allowing us to put Closures to be run before a thread actually goes to sleep for example), rewriting how we start and kill threads (to remove the PTHREAD_CREATE_DETACHED and join on everything, even daemons) or changing how suspension works in some way. Any of these needs much more investigation before we can start working on them. In the meantime it would be good to get as much coverage from our tests as possible.

Another failed exactly once and the final one seems to sometimes cause OOME during class-linking due to allocating a lot of stuff (this one is never run currently).

All of these are "crashes" but in every case either a KI that would likely cause too much slowdown to make worth fixing or is a somewhat expected (though inconvenient) possible outcome of the test.

Patch-set: 4
Reviewer: Gerrit User 1047769 <1047769@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=+1