Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> > Draconian restriction on the reader.
> 
> I don't think it is that draconian.  There will be essentially two memcpy operations per entry (one for the linked list entry, and one for the elf file), and you need to check seqlock after each to ensure the copy succeed.  However, thereafter, you are free to read the copy without any fear (it is a copy, it will not change).

You're assuming that the counter will be regularly rechecked after every copy from the writing process. Can we say that in the comment?

> 
> > let's assume that the seqlock release is right after the load of x into tmp_x.
> 
> I don't follow.  Processing of data is allowed only from the copy and only after the copy has been verified.  The x is loaded from the local copy, and any changes to the linked list will not affect it.

My example assumed that the reader and writer are in the same process, and that x is a shared variable modified by the writer while holding the seqlock for writing, and directly read by the reader while holding a seqlock.

The code below always seems to use a kernel call to read data in the reader. That probably gives you some extra fencing properties, and probably foils a bunch of optimizations on the reader side, since the compiler is unlikely to understand those calls. Does it apply to all readers? If so, that may help a bit.

>  
> > So I think you're saying we have
> 
> Yes, that proved the case where we read S.  For the case where we read S+1:
> 
> Writer:
> W1) Acquire seqlock for writing, reading S, writing S+1
> W1.5) release-fence
> W2) Remove linked list entry A
> W2.5) release-fence
> W3) Release seqlock, writing sequence number S+2
> W3.5) release-fence
> W4) Clobber A
> 
> Reader:
> R1) Read seqlock for reading, reading S+1
> R1.5) acquire-fence
> R2) Read clobbered A
> R2.5) acquire-fence
> R3) Reread seqlock, again getting S+1
> 
> So for S+1: W3.5 synchronizes with R2.5, thus W3 happens before R3. Contradiction.

The reader can easily abort on initially seeing an odd sequence number, so this case isn't very interesting. That's certainly much cheaper than an added fence. I didn't look at it carefully.

> 
> For S+2: W2.5 synchornizes with R1.5, thus W2(delete pointer to A) happens before R2(read pointer to A). Contradiction.

We also have W3 synchronizes with R1, implying that the reader sees all the writers updates to the data structure. If it still reads A, the writer left a dangling pointer. My conclusion:

The R1 read needs to be an acquire read, or you need the R1.5 fence.
You probably want to check for an odd sequence number before reading.
R2.5 is needed.

W1.5 is needed.
W3 needs to have at least release semantics.
W2.5 and W3.5 are not needed.

Note that this produces a nice symmetry between reader and writer code, as you would expect.

> 
> And as you said for S: W1.5 synchronizes with R2.5 and hence W1 happens-before R3. Contradiction.
> 
> So I think it is impossible to read clobbered A, and all of the fences in the CL are needed.
Agreed with the first half, assuming atomic accesses everywhere.
> 
> 
> PS: The WIP reader implementation is here: https://android-review.googlesource.com/#/c/platform/system/core/+/628626/2/libunwindstack/JitDebug.cpp

Thanks! I was trying to formulate restrictions under which this is correct. I think these need to include that all data accesses are either atomic, or "close enough in practice". But there are still things I don't understand. See my comments on the other CL.

Patch-set: 4
Reviewer: Gerrit User 1042828 <1042828@85c56323-6fa9-3386-8a01-6480fb634889>