Update patch set 4
Patch Set 4:

> seems to use a kernel call to read data in the reader

Yes, it is my goal always read the data that way (mainly to protect from SEGV).

> can easily abort on initially seeing an odd sequence number

As mention in the other CL.  That is not an option for crashed or stuck process.

> W3 needs to have at least release semantics. W2.5 and W3.5 are not needed.

W3(release) is equvalent to W2.5+W3(relaxed).  I prefer that, but it is just personal preference.

W3.5 is needed if we support the odd (crashed) seqlocks.

Patch-set: 4
Reviewer: Gerrit User 1065473 <1065473@85c56323-6fa9-3386-8a01-6480fb634889>