Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> > > Patch Set 4: Code-Review-2
>  > >
>  > > While I agree there is value in marking tests as flakes, I fail
>  > > to understand what we're trying to gain by re-trying. Re-trying for
>  > > me only has the potential to load even more our bots for little
>  > > value. (and re-trying could be down to the test itself, it doesn't
>  > > need to be plumbed in the infra).
>  > 
>  > The increased test-infra load will be minimal. It should basically
>  > be just the same as adding a single additional test (in the rare
>  > case that it even runs).
> 
> Sure, but again a new test adds value. Running the same test multiple times I fail to find it.

The value is it allows us to distinguish flakes from actual failures.

>  > 
>  > The added value is that we currently have several tests that we
>  > don't run due to (very rare) flakes. Of the tests I mark as flaky
>  > in go/aog/632215 all of them have single-digit percent flake rates
>  > but currently we don't run them. One of them is never run at all
>  > under any configuration (080-oom-fragmentation). By allowing a
>  > single retry we allow tests to continue to run and to get results.
>  > 
>  > I don't want to put the retry in the test itself since that will
>  > make even worse the huge number of test run-configurations.
> 
> I agree. I think authors of flaky tests should work hard on removing the flakiness. The most deterministic a test can be, the better it will be to try to reproduce an error.

Of course but there can be major difficulties. In the tests I'm marking as flaky the underlying bug is nothing to do with the test but that we don't synchronize daemon thread and runtime shutdown in a safe way (we literally sleep for a few hundred ms and hope that all daemon threads reach a safe spot). This bug has been known for *years* (b/26483935, go/aog/195700) and both in the past and today it's not clear if the effort and perf cost to fix it for real is worth it.

As for changing the tests to reduce flakiness. The hitting the daemon thread issue all require daemon threads to work (they either explicitly test daemon threads or test compiler bugs in code-gen for infinite loops).

>  > Furthermore most tests are only flakey under specific variants
>  > which are most easily checked in the testrunner. By putting it in
>  > the testrunner knownfailures.json it is also easier to remove when
>  > the flakiness is fixed and more visible as long as it's there.
> 
> If they're already being tested in the variant that's meaningful for
> it, it should be ok to just skip it for the variants that lead to it
> being undeterministic.

I think a large part of why we have so many configurations is because we don't always know which variants cause meaningful changes to the runtime. If not then we can get rid of most of the columns in master-art-host.

>  > 
>  > > On the other hand, I do see value in having the test run and
>  > > ensure it doesn't crash, so we don't report a pass/fail, but only
>  > > report a crash. Is that what you're trying to accomplish?
>  > 
>  > This would be very difficult to do because of how we run the tests
>  > (5 layers of shell scripts) and would be insufficient for our
>  > purposes anyway.
>  > 
>  > Of the tests I mark as flaky in go/aog/632215 5 of them fail due to
>  > a rare KI race during shutdown with daemon threads (b/26483935,
>  > b/73871735). In those cases the slight change in performance from
>  > redefine-stress changes the flake rate from ~0.2% to ~0.8% from my
>  > tests. This seems to be just barely enough to (between the 5 tests
>  > with this problem) cause somewhat frequent failures when enabled.
>  > Fixing this in the 'right' way would almost certainly involve
>  > slowing down thread suspension (by allowing us to put Closures to
>  > be run before a thread actually goes to sleep for example),
>  > rewriting how we start and kill threads (to remove the
>  > PTHREAD_CREATE_DETACHED and join on everything, even daemons) or
>  > changing how suspension works in some way. Any of these needs much
>  > more investigation before we can start working on them. In the
>  > meantime it would be good to get as much coverage from our tests as
>  > possible.
> 
> Sure, then would be better to extract the meaningful, non flaky parts of
> a test to one that will always run and pass. And then leave the flake
> part only run on the configuration it's deterministic.

Most of our tests are fairly small already. Also 5 of these flakes are due to a bug caused by the mere use of daemon threads, which are critical to the tests.

>  > 
>  > Another failed exactly once and the final one seems to sometimes
>  > cause OOME during class-linking due to allocating a lot of stuff
>  > (this one is never run currently).
>  > 
>  > All of these are "crashes" but in every case either a KI that would
>  > likely cause too much slowdown to make worth fixing or is a
>  > somewhat expected (though inconvenient) possible outcome of the
>  > test.
> 
> 
> The risk of supporting flakes in the infrastructure itself is that we can
> hide genuine failures, that we would have seen if the test only ran once.
> I feel starting supporting flake at this level is a slippery slope to walk
> away from hard but real bugs (a crash in the wild won't mask itself by re-trying).

We already disable tests if they fail. At least this way we note that the test should generally function and will get a notification if that stops happening.

I feel that what we have now is even worse. We have several times simply disabled a test and then completely forgotten about it. 080-oom-fragmentation has been disabled globally for over a year. I didn't look into every test as I was creating the flake list but I'm sure some of the other skips are similar.

Ensuring that real failures are not ignored or swept under the rug is a challenge that any large team needs to face. It's up to everyone to make sure that features like this are not abused. 

Furthermore with only a single retry granted to a failing test anything that flakes even remotely regularly will still end up red plenty.

Patch-set: 4
Reviewer: Gerrit User 1047769 <1047769@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0