Update patch set 10
Patch Set 10:

> Patch Set 10:
> ...
> 
> The change is nop for low end devices (less than 1GB RAM). On other devices, after 2 seconds, we trigger a GC which should reclaim all the extra memory not in use. The numbers reported shows the maximum heap consumption at the end of the startup, and therefore do not show memory being reclaimed with the change as startup time is less than 2sec in all the cases. The test results are obtained by running "adb shell am start ...; adb shell am force-stop ..."

I missed the 1GB exemption. Good. Does it also apply to exactly 1GB devices?

Most of my concern is with native memory. The heap size is only the Java heap, right? If we collect later, we no doubt have more live native objects associated with those Java objects, so the native heap will grow larger. Since we can't defragment that, we will probably end up with a larger native heap with more fragmentation and more dirty pages that we can't reclaim. I have no idea how big that effect is for these applications. It may not be a big deal, but it would be good to understand it. There are many kinds of Java framework objects that allocate the bulk of their memory requirements in the native heap.

Patch-set: 10
Reviewer: Gerrit User 1042828 <1042828@85c56323-6fa9-3386-8a01-6480fb634889>