Update patch set 10
Patch Set 10:

> Patch Set 10:
> 
> > Patch Set 10:
> > 
> > I missed the 1GB exemption. Good. Does it also apply to exactly 1GB devices?
> > 
> @mathieuc can confirm this. low_memory_mode_ flag in the heap structure is being used to avoid the change on low-memory devices.
> 
> > Most of my concern is with native memory. The heap size is only the Java heap, right? If we collect later, we no doubt have more live native objects associated with those Java objects, so the native heap will grow larger. Since we can't defragment that, we will probably end up with a larger native heap with more fragmentation and more dirty pages that we can't reclaim. I have no idea how big that effect is for these applications. It may not be a big deal, but it would be good to understand it. There are many kinds of Java framework objects that allocate the bulk of their memory requirements in the native heap.
> 
> We temporarily increase the threshold only for java heap. The native allocation threshold is unchanged. Due to that, as you can see in the commit message, for Camera app there are two GCs that took place even with the change. Both the GCs are due to native allocations.

That's likely to mitigate the problem, but not solve it completely. I don't remember immediately how the native thresholds compare to the Java thresholds, but I suspect there are still scenarios in which we end up with a significantly larger native heap this way. This is all complicated by the fact that the GC's understanding of the native heap size is an extremely rough approximation of reality and can sometimes be orders of magnitude too low.

Again, I'm not saying this is a bad idea; I'm just uncomfortable claiming it's a good idea based on the data we currently have. If we could trigger a few collections after the startup interval, and then look at total PSS numbers for the whole process in both scenarios, I would be a lot more comfortable.

Patch-set: 10
Reviewer: Gerrit User 1042828 <1042828@85c56323-6fa9-3386-8a01-6480fb634889>