Add buildbot-logcat.sh script
This script will let the chrome-buildbot filter the logcat output
slightly. This script will pass the output of logcat through
'uniq -f4'. This will combine lines that are identical except for the
date, time, pid or tid they are on. This should reduce the amount of
logging that the buildbot saves while keeping around any important
information.

Bug: 74429665
Test: manual
Change-Id: I66485adff9aa99861ab689e2fb927372d0b01422