Update patch set 4
Patch Set 4:

I am concerned with being able to read state from crashed process, but from live process as well (profiling/debug-malloc). Ideally, I would like to have one algorithm that can read safely regardless whether the process is live, dead, or something in between (mostly to avoid trying to determine the state of the process or adding timeouts). I believe the current algorithm does achieve that.

Even if the counter is odd, there is a lot that can be achieved safely (with this CL, but not before; which is the main point).

1) Read seqlock value 2S+1
2) Read head pointer of the list
3) Memcpy the content of entry
4) Read seqlock and check it is 2S+1
5) Memcpy the symfile from the entry
6) Read seqlock and check it is 2S+1
7) Go to step 3 using the next pointer

The above should result in complete copy of all symfiles which were in the linked list with no data corruption. Or am I missing something?

Patch-set: 4
Reviewer: Gerrit User 1065473 <1065473@85c56323-6fa9-3386-8a01-6480fb634889>