Update patch set 6
Patch Set 6:

(2 comments)

Thanks for the comments.

This is slightly broken at present, but perhaps going in a direction that is easier to read. Until resolved, it's marked as WIP.

I may decouple this from the simplification of atomic.h that currently follows as that is largely mechanical.

Patch-set: 6
Reviewer: Gerrit User 1071150 <1071150@85c56323-6fa9-3386-8a01-6480fb634889>