Give testrunner the ability to collect logcat during tests
It is often useful to have the device logcat from while our tests are
running. Until now this had to be done manually. This adds two new
flags. --logcat-file <file> causes the logcat of the device to be sent
to the given file. --logcat-invoke '<invoke>' lets one control how
the logcat is collected. The output of the command is piped to the
logcat file. The command is killed with SIGTERM when the testing ends.

This is used to ensure that the entirety of the logcat is kept around,
even if there is more than the normal logcat buffer.

Test: ./test/testrunner/testrunner.py \
              --target \
              --logcat-invoke 'adb logcat -v threadtime' \
              --logcat-file /tmp/test-logcat \
              -t 001-HelloWorld
Bug: 74429665

Change-Id: I70527b5b9e1547aa1f97a217ec56fd1ffc0c235c