Update patch set 3
Patch Set 3:

> Looks fine here.
> 
> Should we keep one config with SS? Or add one (possibly in a later CL) if having SS is still useful.

Thanks Orion. Yeah, this is a good point. Thanks for updating the bug as well. Note that we are covering the Semi-Space collector with the following configurations at the moment:
- art-ss-gc
- art-ss-gc-tlab
- art-gtest-ss-gc
- art-gtest-ss-gc-tlab

Maybe we could keep a subset of these.

Patch-set: 3
Reviewer: Gerrit User 1052644 <1052644@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0