Update patch set 5
Patch Set 5:

> Patch Set 4:
> 
> I am concerned with being able to read state from crashed process, but from live process as well (profiling/debug-malloc). Ideally, I would like to have one algorithm that can read safely regardless whether the process is live, dead, or something in between (mostly to avoid trying to determine the state of the process or adding timeouts). I believe the current algorithm does achieve that.
> 
> Even if the counter is odd, there is a lot that can be achieved safely (with this CL, but not before; which is the main point).
> 
> 1) Read seqlock value 2S+1
> 2) Read head pointer of the list
> 3) Memcpy the content of entry
> 4) Read seqlock and check it is 2S+1
> 5) Memcpy the symfile from the entry
> 6) Read seqlock and check it is 2S+1
> 7) Go to step 3 using the next pointer
> 
> The above should result in complete copy of all symfiles which were in the linked list with no data corruption. Or am I missing something?

I think that's probably correct, if you assume the next fields are atomic. But that's a completely separate property that has nothing to do with normal seqlocks; it follows from the fact that there is only a single atomic operation that matters inside the write critical section.

I'm OK with making the next fields atomic and relying on that, assuming this is carefully documented. It only works for very restricted reader behavior. If the reader reads the other descriptor fields, all bets are off.

Patch-set: 5
Reviewer: Gerrit User 1042828 <1042828@85c56323-6fa9-3386-8a01-6480fb634889>